added more comments

+++ app/main/views.py
@@ -8,13 +8,20 @@ from .. import db
 from sqlalchemy.sql.expression import func
 
 def create_question_object(form, question_id=None):
+    """
+    Function to be called in create_question
+    and edit question routes
+
+    return: returns a question instance
+    """
     if question_id:
         question = Question.query.get_or_404(question_id)
     else:
         question = Question()
         #if the question is a new question, assign it a user, which will be the user id of the current user that will definitely be a teacher
         question.user=current_user._get_current_object()
-
+    #The following lines of codes populate the field of the question instance 
+    #with the data submitted by the field and returns it
     question.body = form.body.data
     question.a = form.a.data
     question.b = form.b.data
@@ -29,10 +36,21 @@ def create_question_object(form, question_id=None):
 
 @main.route('/', methods=['GET'])
 def index():
+    """
+    The function that handles the homepage
+
+    return: returns the index html template
+    """
     return render_template('index.html')
 
 @main.route('/courses', methods=['GET', 'POST'])#route for all courses created
 def courses():
+    """
+    This function handles the creation and display of list of available courses
+
+    return: should return a template of list of courses and also a form if for
+    the user to enter a new course name if they are a teacher
+    """
     courses = Course.query.order_by('course_name').all()
     if current_user.is_teacher:
         form = CoursesForm()
@@ -53,6 +71,13 @@ def courses():
 @login_required
 @requires_teacher
 def delete_course(id):
+    """
+    This function handles the deletion of courses  whose ids are passed, it is
+    decorated by the requires_teacher decorator, i.e the user who visited the
+    route that called this function is required to be a teacher
+
+    return: should redirect to the courses route if the deletion was successful
+    """
     course = Course.query.get_or_404(id)
     if course:
         db.session.delete(course)
@@ -62,6 +87,12 @@ def delete_course(id):
 
 @main.route('/course/<int:id>', methods=['GET'])
 def course(id):
+    """
+    This function handles the display of the details of the course that its
+    id was passed to the function
+
+    return: should return a template of the course with its details
+    """
     quizes = Question.query.filter_by(course_id=id)
     if quizes.first() is None:
         flash('There are no questions for this course yet')
@@ -74,12 +105,20 @@ def course(id):
 @login_required
 @requires_teacher
 def create_question():
+    """
+    This function handles the creation of questions
+
+    return: should return a template containing the question form, it is
+    decorated by the requires_teacher decorator, i.e the user who visited the
+    route that called this function is required to be a teacher
+    """
     #sets the default course in the course select field
     form = QuestionForm(course_id=session.get('selected_course'))
     if form.validate_on_submit():
         if not session.get('selected_course') or not form.course_id.data == session.get('selected_course'):
             #if there is no selected course in the session or
-            #if the course that was just selected is not the course that was in the session
+            #if the course that was just selected is not the 
+            #course that was previously in the session,
             #change the course in the session to the one just selected
             session['selected_course'] = form.course_id.data
         
@@ -95,6 +134,14 @@ def create_question():
 @login_required
 @requires_teacher
 def edit_question(id):
+    """
+    This function handles the editing of questions whose ids are passed into
+    the function, decorated by the requires_teacher decorator, also ensures 
+    that the person trying to edit a question is the creator
+
+    return: should return a template containing the question form, the form
+    fields should be filled
+    """
     if not current_user.owns_question(id):
         flash('The question you tried to edit was not created by you')
         return redirect(url_for('main.index'))
@@ -111,8 +158,8 @@ def edit_question(id):
     """
         there is a reason this part is below and not top
         the reason is that, putting it at the top keeps 
-        changing the edited values to initial values 
-        they were before editing
+        changing the edited values to the initial values 
+        that they were before editing
     """
     question = Question.query.get_or_404(id)
     form.body.data = question.body
@@ -130,6 +177,13 @@ def edit_question(id):
 @login_required
 @requires_teacher
 def delete_question(id):
+    """
+    This function handles the deletion of questions whose ids were passed,
+    decorated by the requires_teacher decorator, also it ensures that the
+    person trying to delete the question was the creator
+
+    return: should return a template 
+    """
     if not current_user.owns_question(id):
         flash('The question you tried to delete was not created by you')
         return redirect(url_for('main.index'))
@@ -144,11 +198,23 @@ def delete_question(id):
 @login_required
 @requires_teacher
 def my_questions():
+    """
+    This function handles the display of the currently logged in teacher
+
+    return: should return a template containing the questions created by
+    the currently logged in teacher
+    """
     my_questions = Question.query.filter_by(user=current_user._get_current_object())
     return render_template('my_questions.html', my_questions=my_questions)
 
 @main.route('/quiz', methods=['GET', 'POST'])
 def quiz():
+    """
+    This function handles the quiz
+
+    return: should return a template containing the questions to be answered by the user
+    as well as the result of the just taken test
+    """
     course = request.args.get('course')
     teacher = request.args.get('teacher')
     quizes = Question.query.filter_by(course_id=course).filter_by(user_id=teacher).order_by(func.random()).limit(10)
@@ -191,6 +257,13 @@ def quiz():
 @main.route('/my-results', methods=['GET'])
 @login_required
 def my_results():
+    """
+    This function handles the display of the results of the currently
+    logged in user, decorated by the login_required decorator
+
+    return: should return a template containing all the highest scored
+    to the quizes taken by the currently logged in user
+    """
     my_results = Result.query.filter_by(user=current_user._get_current_object()).all()
     for my_result in my_results:
         my_result.teacher_name = User.query.get_or_404(my_result.teacher_id).username
